+++ backend/server/urls.py
@@ -1,23 +1,24 @@
 from django.urls import path
-from . import views
+# from . import views
+from .views import *
 
 urlpatterns = [
-    path('signin/', views.Login), # POST username and password to receive token and user info
-    path('signup/', views.Register.as_view()), # Sign Up through POST
-    path('kakao/', views.Kakao.as_view()), # Kakao Sign Up or Sign In through POST
-    path('user/', views.GetProfile.as_view()), # Get profile
-    path('user/<int:pk>/', views.ProfileDetail.as_view()), # My page
-    path('user/<int:pk>/meetings/', views.UserMeetingList.as_view()),
-    path('meetings/', views.MeetingList.as_view()),
-    path('meetings/new/<int:id>/', views.RecentMeetingList.as_view()),
-    path('meetings/<int:pk>/', views.MeetingDetail.as_view()),
-    path('meetings/<int:pk>/join/', views.JoinMeeting.as_view()),
-    path('meetings/<int:meeting>/accept/<int:pk>/', views.AcceptMeeting.as_view()),
-    path('meetings/<int:meeting>/reject/<int:pk>/', views.RejectMeeting.as_view()),
+    path('signin/', Login), # POST username and password to receive token and user info
+    path('signup/', Register.as_view()), # Sign Up through POST
+    path('kakao/', Kakao.as_view()), # Kakao Sign Up or Sign In through POST
+    path('user/', GetProfile.as_view()), # Get profile
+    path('user/<int:pk>/', ProfileDetail.as_view()), # My page
+    path('user/<int:pk>/meetings/', UserMeetingList.as_view()),
+    path('meetings/', MeetingList.as_view()),
+    path('meetings/new/<int:id>/', RecentMeetingList.as_view()),
+    path('meetings/<int:pk>/', MeetingDetail.as_view()),
+    path('meetings/<int:pk>/join/', JoinMeeting.as_view()),
+    path('meetings/<int:meeting>/accept/<int:pk>/', AcceptMeeting.as_view()),
+    path('meetings/<int:meeting>/reject/<int:pk>/', RejectMeeting.as_view()),
     # path('meetings/<int:pk>/join/', views) #POST, PUT, DELETE
-    path('search/<str:keyword>/', views.SearchResult.as_view()), #GET
-    path('searchdist/<int:dist>/', views.SearchLocation.as_view()), # GET, search by location, distance
-    path('comment/', views.CommentList.as_view()), # POST, should be here
-    path('comment/<int:pk>/', views.CommentDetail.as_view()), # PUT, DELETE
-    path('notification/<int:pk>/', views.NotificationDetail.as_view()) #GET, DELETE
+    path('search/<str:keyword>/', SearchResult.as_view()), #GET
+    path('searchdist/<int:dist>/', SearchLocation.as_view()), # GET, search by location, distance
+    path('comment/', CommentList.as_view()), # POST, should be here
+    path('comment/<int:pk>/', CommentDetail.as_view()), # PUT, DELETE
+    path('notification/<int:pk>/', NotificationDetail.as_view()) #GET, DELETE
 ]
