+++ backend/server/views/__init__.py
@@ -1,0 +1,6 @@
+from .auth import *
+from .comment import *
+from .meeting import *
+from .notification import *
+from .profile import *
+from .search import *
